Implement retransmission
When packets arrived that are re-transmitted, we evaluate if we need to re-send
whatever we sent last. A positive 'g_pkt_len' indicates we have something to
send. If needed, we send it again.

We track our own state so that we know when to reset retransmission (not
used yet). Also, we rate limit ourselves to be a good wifi-netizen.